Create initial package for PyPI (pip) distribution
